+++ faststack/config.py
@@ -487,6 +487,7 @@ DEFAULT_CONFIG = {
     },
     "raw": {
         "source_dir": "C:\\Users\\alanr\\pictures\\olympus.stack.input.photos",
+        "secondary_source_dir": "H:\\olympus.stack.input.photos",
         "mirror_base": "C:\\Users\\alanr\\Pictures\\Lightroom",
     },
     "updates": {